Add attackspeed to fighting logic add static fightlog to keept track of attackinfos

+++ RPG.Library/Models/Abilities/FireBallAbility.cs
@@ -10,7 +10,7 @@ namespace RPG.Models
 {
     public class FireBallAbility : Ability
     {
-        public FireBallAbility(string name, int damage, int manaCost, int cooldown, StatusEffect statusEffect) : base(name, damage, manaCost, cooldown, statusEffect)
+        public FireBallAbility(string name, int damage, int manaCost, float cooldown, StatusEffect statusEffect) : base(name, damage, manaCost, cooldown, statusEffect)
         {
         }
         public FireBallAbility(FireBallAbility clone) : base(clone.Name, clone.Damage, clone.ManaCost, clone.Cooldown, clone.StatusEffect)
