+++ RPG.Library/Models/Abilities/PoisonAbility.cs
@@ -9,7 +9,7 @@ public class PoisonAbility : Ability
     {
     }
 
-    public PoisonAbility(string name, int damage, int manaCost, int cooldown, StatusEffect statusEffect) : base(name, damage, manaCost, cooldown, statusEffect)
+    public PoisonAbility(string name, int damage, int manaCost, float cooldown, StatusEffect statusEffect) : base(name, damage, manaCost, cooldown, statusEffect)
     {
     }
